Added internationalization
What?
Added internationalization.

Why?
The client of workbench needs it to support different languages.

How?
Update version of ontotext-yasgui-web-component.
