+++ package.json
@@ -105,7 +105,7 @@
         "ng-file-upload": "^12.2.13",
         "ng-tags-input": "^3.2.0",
         "oclazyload": "^1.1.0",
-        "ontotext-yasgui-web-component": "^0.0.2-TR3",
+        "ontotext-yasgui-web-component": "^0.0.2-TR4",
         "shepherd.js": "^10.0.1"
     },
     "resolutions": {
