+++ src/pages/sparql-editor.html
@@ -4,6 +4,7 @@
     <ontotext-yasgui
         ng-custom-element
         ngce-prop-config="config"
+        ngce-prop-language="language"
         ngce-on-query_executed="queryExecuted($event)"
         ngce-on-create_saved_query="createSavedQuery($event)">
     </ontotext-yasgui>
