+++ test-cypress/integration/sparql-editor/internationalization.spec.js
@@ -9,7 +9,6 @@ describe('Internationalization of ontotext-yasgui-web-component', () => {
 
     beforeEach(() => {
         const repositoryId = 'sparql-editor-' + Date.now();
-        cy.intercept('GET', '/rest/monitor/query/count', {body: 0});
         cy.createRepository({id: repositoryId});
         cy.presetRepository(repositoryId);
         QueryStubs.stubDefaultQueryResponse(repositoryId);
