+++ test-cypress/steps/yasgui/yasgui-steps.js
@@ -4,32 +4,6 @@ export class YasguiSteps {
         return cy.get('.yasgui');
     }
 
-    static getYasguiTag() {
-        return cy.get('ontotext-yasgui');
-    }
-
-    static openANewTab() {
-        cy.get('button.addTab').click();
-    }
-
-    static getTabs() {
-        return cy.get('.tab');
-    }
-
-    static getCurrentTab() {
-        return cy.get('.tab.active');
-    }
-
-    static openTab(index) {
-        this.getTabs().eq(index).click();
-    }
-
-    static getTabQuery(tabIndex) {
-        return cy.get('.yasqe .CodeMirror').then((el) => {
-            return el[tabIndex].CodeMirror.getValue();
-        });
-    }
-
     static getCreateSavedQueryButton() {
         return cy.get('.yasqe_createSavedQueryButton');
     }
@@ -38,16 +12,76 @@ export class YasguiSteps {
         this.getCreateSavedQueryButton().click();
     }
 
-    static getShowSavedQueriesButton() {
-        return cy.get('.yasqe_showSavedQueriesButton');
+    static getSaveQueryDialog() {
+        return cy.get('.dialog');
     }
 
-    static showSavedQueries() {
-        this.getShowSavedQueriesButton().click();
+    static closeSaveQueryDialog() {
+        this.getSaveQueryDialog().find('.close-button').click();
+    }
+
+    static cancelSaveQuery() {
+        this.getSaveQueryDialog().find('.cancel-button').click();
+    }
+
+    static getSaveQueryButton() {
+        return this.getSaveQueryDialog().find('.ok-button');
+    }
+
+    static saveQuery() {
+        this.getSaveQueryButton().click();
+    }
+
+    static getQueryField() {
+        return this.getSaveQueryDialog().find('#query');
+    }
+
+    static writeQuery(query) {
+        this.getQueryField().type(query, {parseSpecialCharSequences: false});
+    }
+
+    static clearQueryField() {
+        this.getQueryField().clear();
+    }
+
+    static getQueryNameField() {
+        return this.getSaveQueryDialog().find('#queryName');
+    }
+
+    static writeQueryName(queryName) {
+        this.getQueryNameField().type(queryName);
+    }
+
+    static clearQueryNameField() {
+        this.getQueryNameField().clear();
+    }
+
+    static toggleIsPublic() {
+        this.getSaveQueryDialog().find('#publicQuery').click();
+    }
+
+    static getErrorsPane() {
+        return this.getSaveQueryDialog().find('.alert-danger');
+    }
+
+    static getErrors() {
+        return this.getErrorsPane().find('.error-message');
     }
 
     static getYasguiModeButton() {
         return cy.get('.btn-mode-yasgui');
+    }
+
+    static getYasguiTag() {
+        return cy.get('ontotext-yasgui');
+    }
+
+    static getTabs() {
+        return cy.get('.tab');
+    }
+
+    static openANewTab() {
+        cy.get('button.addTab').click();
     }
 
     static isYasguiModeSelected() {
@@ -144,17 +178,5 @@ export class YasguiSteps {
 
     static isHorizontalOrientation() {
         this.getYasguiTag().should('have.class', 'orientation-horizontal');
-    }
-
-    static getDeleteQueryConfirmation() {
-        return cy.get('.confirmation-dialog');
-    }
-
-    static rejectDeleteOperation() {
-        this.getDeleteQueryConfirmation().find('.cancel-button').click();
-    }
-
-    static confirmDeleteOperation() {
-        this.getDeleteQueryConfirmation().find('.confirm-button').click();
     }
 }
